requirements.txt
# Core dependencies
python-telegram-bot>=20.0
aiosqlite>=0.19.0
PyYAML>=6.0
aiohttp>=3.9.1
aiocron>=1.8

# Security
PyJWT>=2.8.0
cryptography>=41.0.0
passlib>=1.7.4

# API and Web
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
starlette>=0.27.0
python-dotenv>=0.19.0

# Database
asyncpg>=0.29.0
aioredis>=2.0.1
alembic>=1.13.0
SQLAlchemy>=1.4.0

# Statistics and Visualization
matplotlib>=3.8.2
pandas>=2.1.3
seaborn>=0.13.0
plotly>=5.18.0

# Template Engine
Jinja2>=3.1.2

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
aiofiles>=23.2.1
aiologger>=0.7.0
asyncio-throttle>=1.0.2

# Cloud Storage
boto3>=1.28.44

# Development and Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1

# Documentation
Sphinx>=7.2.6
sphinx-rtd-theme>=1.3.0
